parser/requirements.txt
antlr4-python3-runtime==4.11.1
PyYAML==6.0
dataclasses-json==0.5.7
